+++ planx_geostats_provider.py
@@ -14,6 +14,7 @@ from .algorithms.alg_local_moran import LocalMoranAlgorithm
 from .algorithms.alg_bivariate_lee_l import BivariateLeeLAlgorithm
 from .algorithms.alg_spatial_regression import SpatialRegressionAlgorithm
 from .algorithms.alg_spatial_autoregression import SpatialAutoregressionAlgorithm
+from .algorithms.alg_spatial_error_regression import SpatialErrorRegressionAlgorithm
 from .algorithms.alg_global_moran import GlobalMoranAlgorithm
 from .algorithms.alg_incremental_autocorrelation import IncrementalAutocorrelationAlgorithm
 from .algorithms.alg_ripleys_k import RipleysKFunctionAlgorithm
@@ -88,6 +89,7 @@ class PlanXGeoStatsProvider(QgsProcessingProvider):
         self.addAlgorithm(SpatialRegressionAlgorithm())
         self.addAlgorithm(GeneralizedLinearRegressionAlgorithm())
         self.addAlgorithm(SpatialAutoregressionAlgorithm())
+        self.addAlgorithm(SpatialErrorRegressionAlgorithm())
         self.addAlgorithm(ExploratoryRegressionAlgorithm())
         self.addAlgorithm(GWRAlgorithm())
         self.addAlgorithm(MGWRAlgorithm())